Add Kibana autologin feature with basic auth
With this commit, Apache2 will have dedicated location
for user tenant, if user dict will include key: "autologin: basic".
 users:
   - user: "kibana"
     role: "readonly"
     password: "kibana"
     tenant: "test.com"
     autologin: 'basic'

With that, user that will go to https://fqdn.com/test_com_autologin
will be redirected to https://fqdn.com/ which is main location for
Kibana service and they don't need to login with the credentials, because
the header with basicauth was injected.

Change-Id: Ia688377c13111f6fd7a39117bb710f1d2493dd32