Add Index State Management (ISM); remove curator
The Index State Managment (ISM) policy let automatically perform
administrative operation on indices.
Ansible is creating automatically policy that remove indexes older
than 14 days and it is related to index pattern: logstash-<tenant>-*.
Besides that, it creates one policy that is not related with index
pattern (Delete all after 14 days).
NOTE: Only new indexes that will be created after creating the policy
will be associated with it, so to perform cleanup of old indexes, use
"Delete all after 14 days" policy as a administrator in Index Managment
menu available on Opensearch Dashboards or attach existing indices to
the policy.

The ISM policy for deleting old indexes has been crated base on [1][2].

[1] https://opensearch.org/docs/latest/im-plugin/ism/policies/#example-policy
[2] https://docs.aws.amazon.com/opensearch-service/latest/developerguide/ism.html

Change-Id: I9c125516466fdcfa7f45f2d30c0216108c264eda