Add permissions for index pattern ftbfs-{{ user.tenant }}-*
The RDO project would like to shutdown a VM, where there is a
script running, that is generating the required dashboard.
In the new way, Zuul CI will create a container and push the data to the
OpenSearch, so there is no need to keep that VM or even create
artifacts in Zuul and keep it in the logserver, due the OpenSearch
will have it for 14 days (with current configuration).

Change-Id: I8da02b2db6175f63dbf98db4d352effad4ffed1d