Allow kibana viewer user to read all zuul.* index patterns
The Zuul CI got hard coded value how they create a index name
in Elasticsearch or OpenSearch. The name is a combination of:

  zuul.<zuul tenant name>

That raises complication after migration to the GPC where index pattern
is now different than the tenant name. Zuul is not sending any
sensitive data, so there is no reason why the viewer user could not be
able to read all zuul.* index patterns.

Change-Id: I9de8fab8ad313bbdebbfc4101ee2693891de760c